+++ src/location.ts
@@ -1,0 +1,3 @@
+import { ReactLocation } from '@tanstack/react-location';
+
+export const location = new ReactLocation();
